src/projectwordle/requirements.txt
pandas
ipykernel==6.29.5
ipywidgets==8.1.3
spacy==3.7.5
regex==2024.5.15
tqdm==4.66.4
matplotlib==3.9.1
plotly==5.22.0
polars==1.1.0
kedro-datasets==4.0.0
inflect==7.3.1
seaborn==0.13.2
pyarrow==16.1.0
numpy==1.26.4